Separated process for gdb info commands from charmrun and ServThread: charmdebug now creates directly the gdb process, and maintains it separately from ServThread. It also now requires no prefix to every command (previously it required "info:" prepended to each command).
Expanded interaction between AllocationGraph and AllocationTree: AllocationGraph can select single points as well as intervals, and AllocationTree will display correctly given the selection.
To view the AllocationGraph the program does not need to be running. Currently the traces must have been generated before the view is called. Therefore, it is best if the program is *not* running as otherwise it will be overwriting the traces.