BreakPoints: included info in the EpInfo class regarding which processors the BP is set. Setting the text color depending if the BP is set on all selected processors or not
PeSet: new class containing a list of processors. Allow better scalability to high number of processors by aboiding long list of processors. A list of PeSets substitutes the right pe panel. A popup menu is available to inspect sets and create new ones (together with a Dialog box to input the new set in a format similar to projections). The status of the set (running/frozen/dead) is displayed through an icon.

Processor: each processor in the system is represented by an instance of this class. Now the debugger has more information about the exact state each processor is in (running/frozen/dead).

ServThread handles a few more notification messages: breakpoint (enhanced), pe frozen, pe died.