Separate connection process from image view
This change separates CCS management from view code. First, it allows a server
to be created independently of a CCS thread. Then, it splits the image view to
an independent view isolated from CCS, and wraps it in a CCS specific view. This
view has its own CCS thread, and controls for controlling requests. The replies
are forwarded to the image view.

Change-Id: I2dcedb475c58e7dbc5ee934c5227d12383a7a2dd